self_pruning_network/requirements.txt
torch
torchvision
numpy
pandas
matplotlib
seaborn
tqdm
pyyaml
tensorboard
